#1753 moving dialog to core-ui, adding javadocs, ui demo
